+++ bridge-ui/src/hooks/fees/index.ts
@@ -1,6 +1,6 @@
 export { default as useBridgingFee } from "./useBridgingFee";
-export { default as useEthBridgingFee } from "./useEthBridgingFee";
-export { default as useERC20BridgingFee } from "./useERC20BridgingFee";
+export { default as useEthBridgingGasUsed } from "./useEthBridgingGasUsed";
+export { default as useERC20BridgingGasUsed } from "./useERC20BridgingGasUsed";
 export { default as useFeeData } from "./useFeeData";
 export { default as useFees } from "./useFees";
 export { default as useGasFees } from "./useGasFees";
