+++ bridge-ui/src/hooks/fees/useERC20BridgingGasUsed.ts
@@ -1,7 +1,7 @@
 import { Address } from "viem";
 import { useQuery } from "@tanstack/react-query";
 import { BridgeProvider, Chain, ChainLayer, ClaimType, Token } from "@/types";
-import { estimateERC20GasFee, isEth } from "@/utils";
+import { estimateERC20BridgingGasUsed, isEth } from "@/utils";
 
 type UseERC20BridgingFeeProps = {
   account?: Address;
@@ -14,7 +14,7 @@ type UseERC20BridgingFeeProps = {
   claimingType: ClaimType;
 };
 
-const useERC20BridgingFee = ({
+const useERC20BridgingGasUsed = ({
   account,
   token,
   fromChain,
@@ -50,7 +50,7 @@ const useERC20BridgingFee = ({
       !!recipient &&
       claimingType === ClaimType.AUTO_PAID,
     queryFn: async () =>
-      await estimateERC20GasFee({
+      await estimateERC20BridgingGasUsed({
         address: account!,
         recipient,
         token,
@@ -65,4 +65,4 @@ const useERC20BridgingFee = ({
   return { data, isLoading, isError, refetch };
 };
 
-export default useERC20BridgingFee;
+export default useERC20BridgingGasUsed;
